refactor: 拆分持仓管理前后端模块
