+++ app.py
@@ -39,6 +39,7 @@ from goldmonitor import risk_analysis as risk_analysis_core
 from goldmonitor import scheduler as scheduler_core
 from goldmonitor import settings_store as settings_store_core
 from goldmonitor import socket_alert_rules as socket_alert_rules_core
+from goldmonitor import socket_portfolio as socket_portfolio_core
 from goldmonitor import storage_manifest as storage_manifest_core
 from goldmonitor import support_files as support_files_core
 from goldmonitor import targets as targets_core
@@ -6584,199 +6585,30 @@ def on_delete_watch_target(data=None):
     _broadcast_alert_rule_views()
 
 
-@socketio.on("get_portfolio")
-def on_get_portfolio():
-    emit("portfolio_updated", build_portfolio_state())
-
-
-@socketio.on("save_portfolio_position")
-def on_save_portfolio_position(data):
-    try:
-        state = upsert_portfolio_position(data)
-    except ValueError as exc:
-        emit("portfolio_error", {"message": str(exc)})
-        return
-    except OSError:
-        emit("portfolio_error", {"message": "持仓保存失败，请检查配置目录权限。"})
-        return
-    socketio.emit("portfolio_updated", state)
-
-
-@socketio.on("delete_portfolio_position")
-def on_delete_portfolio_position(data=None):
-    position_id = data.get("id") if isinstance(data, dict) else None
-    try:
-        ok, state = delete_portfolio_position(position_id)
-    except OSError:
-        emit("portfolio_error", {"message": "持仓保存失败，请检查配置目录权限。"})
-        return
-    if not ok:
-        emit("portfolio_error", {"message": "未找到持仓记录"})
-        emit("portfolio_updated", state)
-        return
-    _broadcast_alert_rule_views()
-
-
-@socketio.on("save_portfolio_transaction")
-def on_save_portfolio_transaction(data):
-    try:
-        state = upsert_portfolio_transaction(data)
-    except ValueError as exc:
-        emit("portfolio_error", {"message": str(exc)})
-        return
-    except OSError:
-        emit("portfolio_error", {"message": "持仓流水保存失败，请检查配置目录权限。"})
-        return
-    socketio.emit("portfolio_updated", state)
-
-
-@socketio.on("delete_portfolio_transaction")
-def on_delete_portfolio_transaction(data=None):
-    transaction_id = data.get("id") if isinstance(data, dict) else None
-    try:
-        ok, state = delete_portfolio_transaction(transaction_id)
-    except ValueError as exc:
-        emit("portfolio_error", {"message": str(exc)})
-        return
-    except OSError:
-        emit("portfolio_error", {"message": "持仓流水保存失败，请检查配置目录权限。"})
-        return
-    if not ok:
-        emit("portfolio_error", {"message": "未找到持仓流水"})
-        emit("portfolio_updated", state)
-        return
-    socketio.emit("portfolio_updated", state)
-
-
-@socketio.on("preview_import_portfolio_transactions")
-def on_preview_import_portfolio_transactions(data=None):
-    content = data.get("content") if isinstance(data, dict) else ""
-    payload = preview_import_portfolio_transactions_csv(content)
-    request_id = data.get("request_id") if isinstance(data, dict) else ""
-    if request_id:
-        payload["request_id"] = str(request_id)
-    emit("portfolio_import_previewed", payload)
-
-
-@socketio.on("import_portfolio_transactions")
-def on_import_portfolio_transactions(data=None):
-    content = data.get("content") if isinstance(data, dict) else ""
-    try:
-        state, imported_count = import_portfolio_transactions_csv(content)
-    except ValueError as exc:
-        emit("portfolio_error", {"message": str(exc)})
-        return
-    except OSError:
-        emit("portfolio_error", {"message": "持仓流水导入失败，请检查配置目录权限。"})
-        return
-    emit("portfolio_imported", {
-        "ok": True,
-        "kind": "transactions",
-        "count": imported_count,
-        "summary": state.get("import_summary") or {"count": imported_count},
-        "import_backup": state.get("import_backup") or portfolio_import_backup_state(portfolio_import_backup),
-    })
-    socketio.emit("portfolio_updated", state)
-
-
-@socketio.on("undo_portfolio_import")
-def on_undo_portfolio_import():
-    try:
-        ok, state = undo_portfolio_import()
-    except OSError:
-        emit("portfolio_error", {"message": "撤销导入失败，请检查配置目录权限。"})
-        return
-    if not ok:
-        emit("portfolio_import_undo_error", {
-            "ok": False,
-            "kind": "transactions",
-            "code": "no_backup",
-            "message": "没有可撤销的导入批次。",
-            "import_backup": state.get("import_backup") if isinstance(state, dict) else portfolio_import_backup_state(portfolio_import_backup),
-        })
-        emit("portfolio_updated", state)
-        return
-    emit("portfolio_import_undone", {
-        "ok": True,
-        "kind": "transactions",
-        "import_backup": state.get("import_backup") or portfolio_import_backup_state(portfolio_import_backup),
-    })
-    socketio.emit("portfolio_updated", state)
-
-
-@socketio.on("save_portfolio_alert")
-def on_save_portfolio_alert(data):
-    try:
-        state = upsert_portfolio_alert(data)
-    except ValueError as exc:
-        emit("portfolio_error", {"message": str(exc)})
-        return
-    except OSError:
-        emit("portfolio_error", {"message": "持仓提醒保存失败，请检查配置目录权限。"})
-        return
-    _broadcast_alert_rule_views()
-
-
-@socketio.on("reset_portfolio_alert")
-def on_reset_portfolio_alert(data=None):
-    alert_id = data.get("id") if isinstance(data, dict) else None
-    try:
-        ok, state = reset_portfolio_alert(alert_id)
-    except (ValueError, OSError):
-        emit("portfolio_error", {"message": "持仓提醒保存失败，请检查配置目录权限。"})
-        return
-    if not ok:
-        emit("portfolio_error", {"message": "未找到持仓提醒"})
-        emit("portfolio_updated", state)
-        return
-    _broadcast_alert_rule_views()
-
-
-@socketio.on("delete_portfolio_alert")
-def on_delete_portfolio_alert(data=None):
-    alert_id = data.get("id") if isinstance(data, dict) else None
-    try:
-        ok, state = delete_portfolio_alert(alert_id)
-    except OSError:
-        emit("portfolio_error", {"message": "持仓提醒保存失败，请检查配置目录权限。"})
-        return
-    if not ok:
-        emit("portfolio_error", {"message": "未找到持仓提醒"})
-        emit("portfolio_updated", state)
-        return
-    _broadcast_alert_rule_views()
-
-
-@socketio.on("export_portfolio")
-def on_export_portfolio(data=None):
-    kind = data.get("kind") if isinstance(data, dict) else "positions"
-    if kind not in {"positions", "transactions", "review"}:
-        kind = "positions"
-    suffix = kind if kind in {"transactions", "review"} else "positions"
-    extension = "md" if kind == "review" else "csv"
-    filename = f"GoldMonitor-portfolio-{suffix}-{datetime.now().strftime('%Y%m%d-%H%M%S')}.{extension}"
-    try:
-        content, count = build_portfolio_csv(kind)
-        saved_path = save_export_file(filename, content)
-        emit("portfolio_exported", {
-            "ok": True,
-            "kind": kind,
-            "filename": filename,
-            "saved_path": saved_path,
-            "count": count,
-        })
-    except OSError as exc:
-        emit("portfolio_export_error", build_export_error_payload(f"持仓导出失败: {exc}"))
-
-
-@socketio.on("get_portfolio_analytics")
-def on_get_portfolio_analytics(data=None):
-    days = data.get("days") if isinstance(data, dict) else 90
-    try:
-        emit("portfolio_analytics_updated", build_portfolio_analytics_state(days=days))
-    except (ValueError, OSError, sqlite3.Error) as exc:
-        logging.warning("生成持仓分析失败: %s", exc)
-        emit("portfolio_analytics_error", {"message": "持仓收益与预警分析生成失败，请稍后重试。"})
+socket_portfolio_core.register_portfolio_handlers(
+    socketio,
+    build_portfolio_state=lambda: build_portfolio_state(),
+    upsert_portfolio_position=lambda data: upsert_portfolio_position(data),
+    delete_portfolio_position=lambda position_id: delete_portfolio_position(position_id),
+    upsert_portfolio_transaction=lambda data: upsert_portfolio_transaction(data),
+    delete_portfolio_transaction=lambda transaction_id: delete_portfolio_transaction(transaction_id),
+    preview_import_portfolio_transactions_csv=(
+        lambda content: preview_import_portfolio_transactions_csv(content)
+    ),
+    import_portfolio_transactions_csv=lambda content: import_portfolio_transactions_csv(content),
+    undo_portfolio_import=lambda: undo_portfolio_import(),
+    portfolio_import_backup_state=lambda data: portfolio_import_backup_state(data),
+    get_portfolio_import_backup=lambda: portfolio_import_backup,
+    upsert_portfolio_alert=lambda data: upsert_portfolio_alert(data),
+    reset_portfolio_alert=lambda alert_id: reset_portfolio_alert(alert_id),
+    delete_portfolio_alert=lambda alert_id: delete_portfolio_alert(alert_id),
+    broadcast_alert_rule_views=lambda: _broadcast_alert_rule_views(),
+    build_portfolio_csv=lambda kind: build_portfolio_csv(kind),
+    save_export_file=lambda filename, content: save_export_file(filename, content),
+    build_export_error_payload=lambda message: build_export_error_payload(message),
+    build_portfolio_analytics_state=lambda days: build_portfolio_analytics_state(days=days),
+    now_factory=lambda: datetime.now(),
+)
 
 
 @socketio.on("toggle_watch_target")
