+++ templates/index.html
@@ -674,6 +674,7 @@
 <script src="/static/app-shell.js?v={{ app_version }}"></script>
 
 <script src="/static/alert-rule-center.js?v={{ app_version }}"></script>
+<script src="/static/portfolio-center.js?v={{ app_version }}"></script>
 <script src="/static/app.js?v={{ app_version }}"></script>
 </body>
 </html>
